feat: Add logic to store the width and height of the canvas in the struct

+++ wasm/cmd/canvas/structs.go
@@ -3,8 +3,8 @@ package canvas
 import "syscall/js"
 
 type Canvas struct {
-	height  uint8
-	width   uint8
+	height  int
+	width   int
 	element js.Value
 	context js.Value
 }
